+++ app/build.gradle
@@ -34,7 +34,7 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation group: 'com.google.code.gson', name: 'gson', version: '2.7'
     // STEP 1: Add required dependencies
-    implementation 'io.transcend.webview:webview:1.0.0'
+    implementation 'io.transcend.webview:webview:1.0.11-LOCAL7'
     implementation 'androidx.preference:preference:1.2.1'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
