+++ app/src/main/java/io/transcend/samplesdk/MainActivity.java
@@ -11,6 +11,7 @@ import io.transcend.webview.TranscendWebView;
 import io.transcend.webview.models.TrackingConsentDetails;
 
 import android.content.Intent;
+import android.graphics.Color;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
@@ -19,6 +20,7 @@ import android.view.animation.Animation;
 import android.view.animation.ScaleAnimation;
 import android.widget.Button;
 import android.widget.ImageView;
+import android.widget.LinearLayout;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -34,8 +36,6 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        ImageView iconImageView = (ImageView) findViewById(R.id.iconImageView);
-        iconImageView.startAnimation(getZoomOutAnimation());
         setUpTranscendWebView();
         setUpButtons();
     }
@@ -77,6 +77,7 @@ public class MainActivity extends AppCompatActivity {
                     public void onViewReady() {
                         try {
                             transcendInitialized = true;
+                            System.out.println("Transcend Ready!!!!!!!");
                             TranscendAPI.getConsent(getApplicationContext(), trackingConsentDetails -> {
                                 System.out.println("isConfirmed:: " + trackingConsentDetails.isConfirmed());
                                 System.out.println("SharedPreferences: " + PreferenceManager.getDefaultSharedPreferences(getApplicationContext()).getString(TranscendConstants.TRANSCEND_CONSENT_DATA, "lol"));
@@ -86,8 +87,23 @@ public class MainActivity extends AppCompatActivity {
                                     TranscendAPI.getRegimes(getApplicationContext(), regimes -> {
                                         System.out.println("regimes size:: " + regimes.size());
                                         if (regimes.contains("gdpr") && !trackingConsentDetails.isConfirmed()) {
-                                            System.out.println("WebView Alive");
+                                            System.out.println("Requesting user consent...");
                                             transcendWebView.setVisibility(View.VISIBLE);
+                                            transcendWebView.setOnCloseListener(() -> {
+                                                new Handler(Looper.getMainLooper()).post(() -> {
+                                                    try {
+                                                        TranscendAPI.getConsent(getApplicationContext(), consentOnClose -> {
+                                                            System.out.println("In onCloseListener::" + consentOnClose.isConfirmed());
+                                                        });
+                                                    }
+                                                    catch (Exception ex){
+                                                        System.out.println("Exception");
+                                                    }
+                                                });
+                                            });
+                                        } else {
+                                            LinearLayout contentView = findViewById(R.id.contentView);
+                                            contentView.setVisibility(View.VISIBLE);
                                         }
                                     });
                                 } catch (Exception e) {
@@ -96,32 +112,25 @@ public class MainActivity extends AppCompatActivity {
                                 System.out.println(trackingConsentDetails.getPurposes().get("Analytics"));
 
                             });
+
+                            transcendWebView.setOnCloseListener(() -> {
+                                new Handler(Looper.getMainLooper()).post(() -> {
+                                    try {
+                                        TranscendAPI.getConsent(getApplicationContext(), trackingConsentDetails -> {
+                                            System.out.println("In onCloseListener::" + trackingConsentDetails.isConfirmed());
+                                        });
+                                    }
+                                    catch (Exception ex){
+                                        System.out.println("Exception");
+                                    }
+                                });
+                            });
                         } catch (Exception e) {
                             throw new RuntimeException(e);
                         }
                     }
                 }
         );
-    }
-
-    private void getAndLogConsent() {
-        if (transcendInitialized) {
-            try {
-                TranscendAPI.getConsent(getApplicationContext(), new TranscendListener.ConsentListener() {
-                    @Override
-                    public void onConsentReceived(TrackingConsentDetails consentDetails) {
-                        System.out.println("getConsent().isConfirmed(): " + consentDetails.isConfirmed());
-                        System.out.println("getConsent().getPurposes():" + consentDetails.getPurposes());
-                        System.out.println("SharedPreferences: " + androidx.preference.PreferenceManager.getDefaultSharedPreferences(getApplicationContext()).getString(TranscendConstants.TRANSCEND_CONSENT_DATA, "lol"));
-                        System.out.println("GDPR_APPLIES from SharedPreferences: " + PreferenceManager.getDefaultSharedPreferences(getApplicationContext()).getInt(IABConstants.IAB_TCF_GDPR_APPLIES, 100));
-                    }
-                });
-            } catch (Exception e) {
-                System.out.println("getConsent error" + e);
-            }
-        } else {
-            System.out.println("Transcend API not ready!!!");
-        }
     }
 
     public ScaleAnimation getZoomOutAnimation(){
